+++ stream-chat-android-client/src/main/java/io/getstream/chat/android/client/api2/endpoint/MessageApi.kt
@@ -18,10 +18,11 @@ package io.getstream.chat.android.client.api2.endpoint
 
 import io.getstream.chat.android.client.api.AuthenticatedApi
 import io.getstream.chat.android.client.api.QueryParams
-import io.getstream.chat.android.client.api2.model.requests.MessageRequest
 import io.getstream.chat.android.client.api2.model.requests.PartialUpdateMessageRequest
 import io.getstream.chat.android.client.api2.model.requests.ReactionRequest
 import io.getstream.chat.android.client.api2.model.requests.SendActionRequest
+import io.getstream.chat.android.client.api2.model.requests.SendMessageRequest
+import io.getstream.chat.android.client.api2.model.requests.UpdateMessageRequest
 import io.getstream.chat.android.client.api2.model.response.MessageResponse
 import io.getstream.chat.android.client.api2.model.response.MessagesResponse
 import io.getstream.chat.android.client.api2.model.response.ReactionResponse
@@ -40,22 +41,31 @@ import retrofit2.http.Query
 @AuthenticatedApi
 internal interface MessageApi {
 
+    /**
+     * [REST documentation](https://getstream.io/chat/docs/rest/#messages-sendmessage)
+     */
     @POST("/channels/{type}/{id}/message")
     fun sendMessage(
         @Path("type") channelType: String,
         @Path("id") channelId: String,
-        @Body message: MessageRequest,
+        @Body message: SendMessageRequest,
     ): RetrofitCall<MessageResponse>
 
     @GET("/messages/{id}")
     fun getMessage(@Path("id") messageId: String): RetrofitCall<MessageResponse>
 
+    /**
+     * [REST documentation]()https://getstream.io/chat/docs/rest/#messages-updatemessage)
+     */
     @POST("/messages/{id}")
     fun updateMessage(
         @Path("id") messageId: String,
-        @Body message: MessageRequest,
+        @Body message: UpdateMessageRequest,
     ): RetrofitCall<MessageResponse>
 
+    /**
+     * [Rest documentation](https://getstream.io/chat/docs/rest/#messages-updatemessagepartial-request)
+     */
     @PUT("/messages/{id}")
     fun partialUpdateMessage(
         @Path("id") messageId: String,