+++ stream-chat-android-client/src/main/java/io/getstream/chat/android/client/api2/model/requests/SendMessageRequest.kt
@@ -1,5 +1,5 @@
 /*
- * Copyright (c) 2014-2022 Stream.io Inc. All rights reserved.
+ * Copyright (c) 2014-2023 Stream.io Inc. All rights reserved.
  *
  * Licensed under the Stream License;
  * you may not use this file except in compliance with the License.
@@ -18,19 +18,21 @@ package io.getstream.chat.android.client.api2.model.requests
 
 import com.squareup.moshi.JsonClass
 import io.getstream.chat.android.client.api2.endpoint.MessageApi
+import io.getstream.chat.android.client.api2.model.dto.UpstreamMessageDto
 
 /**
- * Used to form a partial message update request.
- * @see [MessageApi.partialUpdateMessage]
+ * Used to form a send message request.
+ * @see [MessageApi.sendMessage]
  *
- * @param set Sets new field values.
- * @param unset Array of field names to unset.
+ * @param message The upstream version of the message.
+ * @param skip_push If the message should skip triggering a push notification when sent. False by default.
  * @param skip_enrich_url If the message should skip enriching the URL. If URl is not enriched, it will not be
  * displayed as a link attachment. False by default.
  */
+// REST documentation: https://getstream.io/chat/docs/rest/#messages-sendmessage
 @JsonClass(generateAdapter = true)
-internal data class PartialUpdateMessageRequest(
-    val set: Map<String, Any>,
-    val unset: List<String>,
+internal data class SendMessageRequest(
+    val message: UpstreamMessageDto,
+    val skip_push: Boolean = false,
     val skip_enrich_url: Boolean = false,
 )